NormalButton -> RectangleButton로 변경
