+++ common/src/main/res/drawable/gray_rectangle_button_drawable.xml
