+++ common/src/main/res/drawable/red_rectangle_button_drawable.xml
