+++ common/src/main/res/layout/layout_rectangle_button.xml
@@ -8,10 +8,10 @@
         android:layout_height="match_parent">
 
         <androidx.appcompat.widget.AppCompatButton
-            android:id="@+id/normal_button_body"
+            android:id="@+id/rectangle_button_body"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:background="@drawable/red_button_drawable"
+            android:background="@drawable/red_rectangle_button_drawable"
             android:textColor="@color/red_button_text_color"
             android:textSize="18sp"
             tools:text="Button"
